requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
fastapi==0.133.1
filelock==3.24.3
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
mangum==0.21.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pluggy==1.6.0
postgrest==2.28.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.28.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0